refactor filter module
