+++ docs/modules/pyfar.dsp.filter.rst
@@ -1,6 +1,12 @@
 pyfar.dsp.filter
 ================
 
+The following documents the pyfar filters. Visit
+:py:mod:`filter types <pyfar._concepts.filter_types>` for an introduction of
+the different filters and
+:py:mod:`filter classes <pyfar._concepts.filter_classes>` for more information
+on pyfar filter objects.
+
 .. automodule:: pyfar.dsp.filter
    :members:
    :undoc-members:
