+++ pyfar/dsp/filter/__init__.py
@@ -1,0 +1,40 @@
+from .classic import (
+    butterworth,
+    chebyshev1,
+    chebyshev2,
+    elliptic,
+    bessel,
+    crossover
+)
+
+from .audiofilter import (
+    bell,
+    high_shelve,
+    low_shelve,
+    high_shelve_cascade,
+    low_shelve_cascade
+)
+
+from .fractional_octaves import (
+    fractional_octave_bands,
+    reconstructing_fractional_octave_bands,
+    fractional_octave_frequencies
+)
+
+
+__all__ = [
+    'butterworth',
+    'chebyshev1',
+    'chebyshev2',
+    'elliptic',
+    'bessel',
+    'crossover',
+    'bell',
+    'high_shelve',
+    'low_shelve',
+    'high_shelve_cascade',
+    'low_shelve_cascade',
+    'fractional_octave_bands',
+    'reconstructing_fractional_octave_bands',
+    'fractional_octave_frequencies'
+]
