+++ pyfar/dsp/filter/_audiofilter.py
